+++ app/components/home/HomeKit.tsx
@@ -1,4 +1,10 @@
-import { Grid, BookOpen, FileText, CogFour } from '@mynaui/icons-react'
+import {
+  Home,
+  BookOpen,
+  FileText,
+  CogFour,
+  InfoCircle,
+} from '@mynaui/icons-react'
 import { ItoIcon } from '../icons/ItoIcon'
 import { useMainStore } from '@/app/store/useMainStore'
 import { useEffect, useState } from 'react'
@@ -7,6 +13,7 @@ import HomeContent from './contents/HomeContent'
 import DictionaryContent from './contents/DictionaryContent'
 import NotesContent from './contents/NotesContent'
 import SettingsContent from './contents/SettingsContent'
+import AboutContent from './contents/AboutContent'
 
 export default function HomeKit() {
   const { navExpanded, currentPage, setCurrentPage } = useMainStore()
@@ -39,16 +46,18 @@ export default function HomeKit() {
         return <NotesContent />
       case 'settings':
         return <SettingsContent />
+      case 'about':
+        return <AboutContent />
       default:
         return <HomeContent />
     }
   }
 
   return (
-    <div className="flex h-full bg-slate-50">
+    <div className="flex h-full">
       {/* Sidebar */}
       <div
-        className={`${navExpanded ? 'w-64' : 'w-20'} flex flex-col justify-between py-4 px-4 transition-all duration-100 ease-in-out`}
+        className={`${navExpanded ? 'w-48' : 'w-20'} flex flex-col justify-between py-4 px-4 transition-all duration-100 ease-in-out border-r border-neutral-200`}
       >
         <div>
           {/* Logo and Plan */}
@@ -57,13 +66,13 @@ export default function HomeKit() {
             <span
               className={`text-2xl font-bold transition-opacity duration-100 ${showText ? 'opacity-100' : 'opacity-0'} ${showText ? 'ml-2' : 'w-0 overflow-hidden'}`}
             >
-              Ito
+              ito
             </span>
           </div>
           {/* Nav */}
           <div className="flex flex-col gap-1 text-sm">
             <NavItem
-              icon={<Grid className="w-5 h-5" />}
+              icon={<Home className="w-5 h-5" />}
               label="Home"
               isActive={currentPage === 'home'}
               showText={showText}
@@ -90,12 +99,19 @@ export default function HomeKit() {
               showText={showText}
               onClick={() => setCurrentPage('settings')}
             />
+            <NavItem
+              icon={<InfoCircle className="w-5 h-5" />}
+              label="About"
+              isActive={currentPage === 'about'}
+              showText={showText}
+              onClick={() => setCurrentPage('about')}
+            />
           </div>
         </div>
       </div>
 
       {/* Main Content */}
-      <div className="flex flex-col w-full items-center bg-white rounded-lg m-2 ml-0 mt-0 border border-neutral-200 pt-12">
+      <div className="flex flex-col flex-1 items-center bg-white rounded-lg m-2 ml-0 mt-0 pt-12">
         {renderContent()}
       </div>
     </div>